u-boot-2009.03-p2020rdb-eSDHC-IRQ-Bypass-Workaround
AHB2MAG IRQ Bypass Hardware Workaround

This is a workaround for the hardware bug found on the P2020 Rev 1.0.
The DCR[DMA__AHB2MAG_IRQ_BYPASS]is not set automatically upon SoC reset.
This patch sets the bit.

Known Issue: mmcinfo has to be run at the u-boot prompt for this patch code to run.
TODO: Move this code to platform initialization code.

Signed-off-by: Vishnu Suresh <Vishnu@freescale.com>