Set HID1[mbdd] bit for optimal eieio
This patch checks the version number and revision number
from the pvr register and sets the HID1[mbdd] bit accordingly.

HID1[mbdd] is found on the new revisions of the e500 cores
which will optimize eieio instruction.

By setting this bit, a 10% improvement is seen in
applications like IP forwarding.

Signed-off-by: Sandeep Gopalpet <sandeep.kumar@freescale.com>