Multiblock Support for MMC
Added Multi-Block Read support for MMC. Modified existing multi-block write
to limit the maximum number of blocks per transfer.  This feature is enabled
with CONFIG_MMC_MBLOCK option.  A new member is added in the mmc structure
for the host controllers to specify the maximum number of blocks it supports.
If the feature is disabled all read/write transfers are single-block.

Signed-off-by: Alagu Sankar <alagusankar@embwise.com>