da850/omapl138: Correct "mmc rescan" implementation
Earlier "mmc rescan" command from U-Boot prompt would always
return ZERO, hence it was not possible to use this command to
know whether the MMC/SD card is present in the slot or not.
This patch corrects this issue.

In the process, modify the bootargs to include a check to find
whether the MMC/SD card is present or not.

Signed-off-by: Sudhakar Rajashekhara <sudhakar.raj@ti.com>