Add microsecond boot time measurement
This defines the basics of a new boot time measurement feature. This allows
logging of very accurate time measurements as the boot proceeds, by using
an available microsecond counter.

To enable the feature, define CONFIG_BOOTSTAGE in your board config file.
Also available is CONFIG_BOOTSTAGE_REPORT which will cause a report to be
printed just before handing off to the OS.

BUG=chromium-os:13875
TEST=build and boot, check that progress is reported before running Linux:

Timer summary in microseconds:
       Mark    Elapsed  Stage
          0          0  awake
  2,181,078  2,181,078  usb_start
 11,861,817  9,680,739  bootp_start
 11,884,610     22,793  bootp_stop
 11,884,689         79  tftp start
 15,271,536  3,386,847  tftp done
 15,271,568         32  bootm_start
 15,406,551    134,983  start_kernel

Change-Id: I71b89c8402dc5dec75e68333bd24a6bab7500a1b
Reviewed-on: http://gerrit.chromium.org/gerrit/197
Reviewed-by: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>