arm: tegra2: Simplify boot path
The Tegra2 boot path is more complicated than it needs to be. Since we want
to move to building most of U-Boot with ARMv7 and only a small part with
ARMv4T (for AVP) it should be as simple as possible.

This makes tegra2_start() into a simple function which either does AVP
init or A9 init depending on which core is running it.

Only two files need to be built for ARMv4T, and this is handled in the
Tegra2 CPU Makefile.

BUG=chromium-os:13875
TEST=Boot U-Boot on Seaboard and run kernel

Change-Id: If07628bd8227c94a2b0f11f2ba9cd71ab41336b2
Reviewed-on: http://gerrit.chromium.org/gerrit/613
Reviewed-by: Tom Warren <twarren@nvidia.com>
Reviewed-by: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>