mmc: Tegra2: Support DMA restarts at buffer boundaries
Currently if a DMA buffer straddles a buffer alignment boundary
(512KiB) then the DMA engine will pause and generate a DMA
interrupt.  Since the DMA interrupt is not enabled it will hang
the MMC driver.

This patch adds support for restarting the DMA transfer.  The
SYSTEM_ADDRESS register contains the next address that would have
been read/written when a boundary is hit.  So we can read that
and write it back.  The write triggers the resumption of the
transfer.

BUG=None
TEST=Run "mmc part 0" and "ext2ls mmc 0:3" on Seaboard.

Signed-off-by: Anton Staaf <robotboy@chromium.org>
Change-Id: Iaa7c37dbeb46e8ae987c7eaf1dcb4dd810dd4160
Reviewed-on: http://gerrit.chromium.org/gerrit/2111