ext2: Simplify partial sector access logic
Previously reading or writing zero full sectors (reading the end of
one sector and the beginning of the next for example) was special
cased and involved stack allocating a second sector buffer.  This
change uses the same code path for this case as well as when there
are a non-zero number of full sectors to access.  The result is
easier to read and reduces the maximum stack used.

BUG=None
TEST=run "ext2ls mmc 0:3" on Seaboard

Change-Id: Iac92172bf40724302292a2a9695cb61029b02ae7
Reviewed-on: http://gerrit.chromium.org/gerrit/2327
Reviewed-by: Anton Staaf <robotboy@chromium.org>
Tested-by: Anton Staaf <robotboy@chromium.org>