Fix device tree compilation for out of tree builds.
The device tree compiler output should be placed in the $(obj)
rooted tree. Since $(OBJCOPY) insists on adding the path to the
generated symbol names, to ensure consistency it should be
invoked from the directory where the .dtb file is located and
given the input file name without the path.

BUG=chromium-os:13082
TEST=build the image for seaboard out of tree (using the
 O=/tmp/dir make parameter)

Change-Id: I8aabae757072c937a42c4dd85e43fdc1309f3dc3
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/2371
Reviewed-by: Simon Glass <sjg@chromium.org>