fdt: tegra2: Add MMC ports
This defines the MMC ports on Seaboard. The first port is MMC4, used for
the internal eMMC device. The second port, MMC3, is used for the SD card.

The peripheral IDs are added to the tegra250 file since they are generic.
These come from the periph_id enum in clock.h. They allow each peripheral
to be associated with its clock, reset and pinmux logic.

BUG=chromium-os:834
TEST=Build and boot u-boot; run mmc_boot
Check that MMC can be accessed still.
Use 'mmc part 0' and 'mmc part 1' to check that both MMC cards work

Change-Id: Ic8b2749458d54bbbfa4143eca42d0ba27401fb08
Reviewed-on: http://gerrit.chromium.org/gerrit/2780
Reviewed-by: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>