Allow pick device tree by specifying an environment variable.
The default device tree used by the make system is
board/nvidia/seaboard/tegra2-seaboard.dts, it is expected to be
defined in the main config file used by the related targets. If
'DEV_TREE_SRC' is defined in the environment, the file
board/nvidia/seaboard/${DEV_TREE_SRC} is used as the device tree
source instead. ./mkconfig is being restored to remove the device
tree related changes.

BUG=chromium-os:834
TEST=manual

Change-Id: I2232c5f4413b719057ec3def77ca03a78068c3c9
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/2776
Reviewed-by: Simon Glass <sjg@chromium.org>