Add support for booting Linux using the 32 bit boot protocol.
This change conditionally modifies the zboot command so that it can use the 32
bit boot protocol. This is necessary because the 16 bit realmode entry point
assumes that it can call BIOS services which neither coreboot nor u-boot
provide.

BUG=chrome-os-partner:4700
TEST=Built and booted using the legacy and 32 bit boot protocols. Used dmesg to
verify that the e820 memory map was successfully picked up by the kernel when
using the 32 bit protocol.

Change-Id: Ibd7d0989a2bef16affb1c3dd268ef9a72461597f
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: http://gerrit.chromium.org/gerrit/3220
Reviewed-by: Stefan Reinauer <reinauer@google.com>
Reviewed-by: Simon Glass <sjg@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>