fdt: update fdt before passing it to kernel
A new flag CONFIG_OF_UPDATE_FDT_BEFORE_BOOT is added to control whether
the device tree is updated before passing it to kernel.

The (partial) calling sequence is:

 do_bootm()
 |
 +--> bootm_linux_fdt()
      |
      +--> fit_update_fdt_before_boot()
      |
      +--> boot_relocate_fdt()
      |
      +--> fdt_chosen()

The fit_update_fdt_before_boot() is called before boot_relocate_fdt()
because the device tree could be enlarged after updating, and so it has
to be called earlier than relocation (otherwise it might not boot).

BUG=chromium-os:15744
TEST=make

Change-Id: Id2abe13e177ac1561f5c5e035acc5e115dba5640
Reviewed-on: http://gerrit.chromium.org/gerrit/3625
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>
Tested-by: Che-Liang Chiou <clchiou@chromium.org>