arm: tegra: fix modifier+key detection
When the user presses Ctrl+D, the top of the keyboard controller FIFO is usually
filled with several "Ctrl" only keycodes. The modifier keys handling code
would remove the keycode and return 0. So, the "tstc" will indicate
that no key is pressed.

The correct behaviour for "tstc" is to check all the entries of the FIFO until
we find a significant one or return 0 if there is none.

BUG=chromium-os:17781
TEST=On Aebl, when the developer screen is displayed, check that the
first Ctrl+D key press is taken into account.

Change-Id: I0d1b4b8928951bac2fc05730083d956642ff4d3e
Reviewed-on: http://gerrit.chromium.org/gerrit/4327
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Anton Staaf <robotboy@chromium.org>