CHROMIUM: use a fixed binary layout of crossystem data
The twostop design is envolved that crossysmem data is now not
reinitialized again in readwrite main firmware, and relied on bootstub
passing the blob around.

Thus, the blob should have a fixed binary layout so that future
readwrite main firmware can reliably parse the blob contents. Currently
the readwrite main firmware just check the sanity of the blob.

BUG=chromium-os:17876
TEST=crossystem blob passes the sanity check by readwrite main firmware

Change-Id: I2981c10234c20d34e1405b5ab127bdc90faa2e3d
Reviewed-on: http://gerrit.chromium.org/gerrit/4373
Tested-by: Che-Liang Chiou <clchiou@chromium.org>
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>