Only define the static secure_boot_cmd function if it's going to be used
The secure_boot_cmd was added to main.c recently and was compiled
unconditionally but used conditionally. gcc warns if a static function is
defined but not used, that was turned into an error, and the build would fail
depending on the configuration options chosen. This change makes both the
definition and use conditionally compiled.

BUG=none
TEST=Build for x86-alex (which compiles it out) and tegra2_kaen (which
compiles it in).

Change-Id: I552eb057b8e1e4cba5d1481b0dc9a7f876a4e81b
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: http://gerrit.chromium.org/gerrit/4954
Tested-by: Gabe Black <gabeblack@chromium.org>
Reviewed-by: Doug Anderson <dianders@chromium.org>