Tegra: pmu: Set SM1 in PWM-only mode
If the Tps6586x SM1 is operating in the ‘AutoPFM/PWM’ mode,
the output may overshoot a couple of hundred millivolts followed by
it undershooting by a couple of hundred millivolts.
Set SM1 operate in PWM-only mode can fix this issue.
The Kernel codes has the same fix, u-boot also should include it.

BUG=None
TEST=Under a very specific load, frequency and output level,
the problem can be observed. So it's difficult to verify.
The kernel codes has the same fix in
http://codereview.chromium.org/6792068

Change-Id: Ieb8dbd4aba4ea1cc7538b631e5defb23df0e7b4e
Signed-off-by: Wei Ni <wni@nvidia.com>
Reviewed-on: http://gerrit.chromium.org/gerrit/5066
Reviewed-by: Yen Lin <yelin@nvidia.com>
Reviewed-by: Jimmy Zhang <jimmzhang@nvidia.com>
Reviewed-by: Doug Anderson <dianders@chromium.org>