Tegra: pmu: Fix the pmu_read and pmu_write
pmu_read/pmu_write call the i2c_read/i2c_write to read/write.
In the old codes, the parameter "alen" is 0, but according to the
tegra-i2c.c, if the alen==0, the i2c_read/i2c_write will be failed
with i2c_addr_ok(), it means the alen should not be 0.
It should be 1 in here.

BUG=chromium-os:18648
TEST=test on seaboard, the pmu_read/pmu_write works fine.

Change-Id: Ib55811bcc259a9dc74993728a23329abc6f2fe5d
Signed-off-by: Wei Ni <wni@nvidia.com>
Reviewed-on: http://gerrit.chromium.org/gerrit/5206
Reviewed-by: Doug Anderson <dianders@chromium.org>
Tested-by: Doug Anderson <dianders@chromium.org>