arm: Tegra2: gpio: Disable pullups for state switches
The write protect switch has a 200K inline resistor that is weaker
than the internal Tegra pull-up.  The result is that even when the
write protect switch is pulling the write protect line to ground
the tegra can't tell.  This doesn't effect the ability of the
write protect switch to write protect the boot flash.  It just
prevents the Tegra from reading the state of the switch.  The
solution is to disable internal pull-ups for the ATD pingroup.  The
ATD pinground is four pins, three of them are used.  The SDIO3 write
protect switch and google recovery mode switch are also wired to the
ATD pingroup.  Both of these switches have external pull-ups, making
it safe to disable the internal ones.

BUG=chrome-os-partner:5040
TEST=verify from U-Boot prompt that Write Protect switch can be read:

gpio info 59
gpio input 59

<toggle WP switch>

gpio input 59

<observe different read value>

Change-Id: If1b8da3a0d8e34a6a997c9f2fa9064f71c93de39
Reviewed-on: http://gerrit.chromium.org/gerrit/5230
Tested-by: Anton Staaf <robotboy@chromium.org>
Reviewed-by: Doug Anderson <dianders@chromium.org>