Add an ICH SPI driver based on the one in flashrom
This change adds a new ICH SPI driver based heavily on the one in flashrom,
and then refactored in a few places. A few functions which would control
aspects of the bus that are handled transparently in hardware simply return
right away.

BUG=chrome-os-partner:4722
TEST=Built on x86-alex, and also built with other changes which allowed reading
the ID from the flash chip. Read the "LARCHIVE" string from the beginning of
u-boot.

Signed-off-by: Gabe Black <gabeblack@google.com>
Change-Id: Ic768c9479523c8392e1197a69a339539e892a9c2
Reviewed-on: http://gerrit.chromium.org/gerrit/5377