Update pci_ids.h from our current Linux sources.
The LPC bus bridge for ICH7 was missing from the pci_ids.h in u-boot. Rather
than add just that device, this change updates the entire file so that we'll
be less likely to run into another ID that's too new.

This change copies over the pci_ids.h file from Linux verbatim. The last
non-merge change hash in that file was:

8930c8aa740b12ad69f44a35137bcc39bfa3dc41

and the kernel was at version 2.6.38.

BUG=chrome-os-partner:4722
TEST=Built x86-alex and tegra2_kaen.

Change-Id: Ibae9361dbf63c278c275f0a8977c31d3470320aa
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: http://gerrit.chromium.org/gerrit/5376
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>