Enable OF_CONTROL for Alex.
This change makes the Alex u-boot use configuration settings
coming from the device tree supplied by coreboot through the
coreboot table.

The device tree now includes the necessary console configuration
information.

Code is added to retrieve the device tree pointer from the
coreboot table, the appropriate coreboot structures and tag
values are being duplicated in
arch/i386/include/asm/ic/coreboot/tables.h. The pointer to the
FDT is added to the global data structure, it is initialized as
soon as it is retrieved from the coreboot table.

For some reason placing the global data structure anywhere but in
the .data segment causes the system to crash. This phenomenon
will be investigated shortly, this commit is an intermediate step
to get the device tree handover used by other coreboot/u-boot
contributors.

Core retrieving the device tree from CBFS is not needed anymore
and gets removed.

BUG=chrome-os-partner:5248
TEST=manual
 . build the new image and program it on Alex
 . restart the target, observe the console output

The target boots all the way to u-boot prompt.

Change-Id: I087cf70362ede0873be37fa5a98aa66a0b979d8f
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/6132
Reviewed-by: Stefan Reinauer <reinauer@google.com>