Fix compiler warnings generated by gcc 4.6.
The newer compiler version is more thorough in detecting code
inconsistencies, and reports warnings in many cases when building
u-boot for kaen and alex.

This change modifies the code to get rid of the warnings. There
supposed to be no logical changes, so no testing other than
building the system is being done.

The files not yet upstreamed are excluded and will be submitted
separately.

BUG=chromium-os:18862
TEST=manual
 . run the following commands

   emerge-tegra2_kaen chromeos-u-boot
   emerge-x86-alex chromeos-u-boot

 . observe them succeed

Change-Id: I4c872d84352539d24a418ba910274d08d02d26a8
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/5706
Reviewed-by: Simon Glass <sjg@chromium.org>