Remove explicit support for the OXPCIE952
Early in development, we needed a driver for the OXPCIE952 UART which would be
left configured by coreboot and just needed to send and receive characters. It
was only partially developed and was somewhat redundant with the NS16550 UART
support already in u-boot. Now that we're using the device tree to configure
the serial console and it requires us to use the NS16550 driver, the OXPCIE952
driver is no longer necessary. Since it isn't likely to be useful in other
contexts, this change removes support for it entirely.

BUG=chrome-os-partner:4520
TEST=Built on x86-alex and tegra2_kaen. Installed and booted on x86-alex and
verified that the serial console still worked.

Change-Id: I124c4596b224f1b2a8d27b2b38e08ee2ce739c79
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: http://gerrit.chromium.org/gerrit/5871
Reviewed-by: Simon Glass <sjg@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>