Expand flash map to maximize size for U-Boot
This increases the size of RO U-Boot to 1MiB and RW U-Boot to 950KiB
to maximize flexibility in the future. We cannot change the FMAP later
so this is trying to cover for all eventualities.

Note: This commit does not affect Seaboard, since it only has 2MiB
of SPI flash.

BUG=chromium-os:19238
TEST=build for Aebl, Kaen, Seaboard

Change-Id: Ifb57fec75eba06eb9946e6a8d7e8100bd5e92c57
Reviewed-on: http://gerrit.chromium.org/gerrit/6227
Reviewed-by: Randall Spangler <rspangler@chromium.org>
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>