Automatically build the device tree if DEV_TREE_SEPARATE is set
Part 3 of 3 of a change to make u-boot build the device tree instead of a
separate ebuild.

This change makes the u-boot "all" target build the separate device tree if
DEV_TREE_SEPARATE is set. This use of DEV_TREE_SEPARATE is consistent with the
one other use in the u-boot build system. It's at least expedient and perhaps
necessary to build the device tree this way since it's difficult to know the
value of $(obj) when invoking make.

TEST=Built and booted on x86-alex, ran vboot_twostop.
BUG=chrome-os-partner:5664

Change-Id: Icdcda86421e6f3cde59439b2be001ff522d6a6e9
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: http://gerrit.chromium.org/gerrit/6460
Reviewed-by: Simon Glass <sjg@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>