CHROMIUM: disk: part_efi: fix **pgpt_pte == NULL
Code was setting **pgpt_pte == NULL, which meant that the pointer
to the gpt_pte would be stored at RAM address 00000000. This 'worked'
on T20 (SDRAM starts @ 0x00000000), but hung gpt/EFI access on T30
(SDRAM starts @ 0x80000000).

NOTE: this is based on an original change from nVidia (see Original-
lines below), but was modified to remove the unnecessary use of a
global variable.

BUG=chromium-os:21540
TEST=Can now run 'mmc rescan 1' and 'mmc part 1' on T30 board
(with WIP t30 patches applied).
TEST=No can still do mmc commands on Kaen.

Change-Id: I0e70bbde3a130b93bf3dd862e1330b41088dd0b0
Original-Change-Id: Idf877d17753b7f4da0180f342276ac28249e1980
Original-Reviewed-on: http://git-master/r/49880
Original-Tested-by: Tom Warren <twarren@nvidia.com>
Original-Reviewed-by: Tom Warren <twarren@nvidia.com>
Signed-off-by: Tom Warren <twarren@nvidia.com>
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/10271
Reviewed-by: Simon Glass <sjg@chromium.org>