Modify the ICH driver to support ICH9 in software sequencing mode
This change modifies the existing ICH SPI controller driver so that it
continues to support the ICH7 controller, but now also the ICH9 controller.
This implementation uses software sequencing mode for two reasons. First, it's
very similar to the only mode of the ICH7 and makes supporting both controllers
much easier. Second, it's much more flexible and actually fits into the u-boot
driver API.

BUG=chrome-os-partner:5829
TEST=With this and other changes, built and vbooted on x86-alex and stumpy.

Signed-off-by: Gabe Black <gabeblack@google.com>

Change-Id: Ifafd5febab8c905760c7e58eca82ee8a383efe36
Reviewed-on: http://gerrit.chromium.org/gerrit/7163
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Gabe Black (Do Not Use) <gabeblack@google.com>