Fix AHCI driver to operate on Intel controllers.
First - make sure the ahci driver compiles and links properly when
building for x86 platforms.

Then it turned out that when trying to use AHCI driver on an x86
platform with an Intel AHCI controller, the driver does not operate
properly if the requested amount of blocks to read was exceeding 255.

It is probably possible to specify 0 as the block count and the driver
will read 256 blocks, but it was decided to limit the number of blocks
read at once to 128 (it should be a power of 2 for the optimal
performance of solid state drives).

BUG=chromium-os:19837
TEST=manual

 . program updated image (including vbexport SATA support extension)
   on an Alex.
 . restart the machine and enter `vboot_twostop' at u-boot prompt.

Observe ChromeOS booting all the way to login screen.

Change-Id: I7224ca14ae60f414db6dbe9e2f0a649312a9459c
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/7232
Reviewed-by: Gabe Black (Do Not Use) <gabeblack@google.com>
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>