Make SATA default storage type on ChromeOS x86 systems.
This adds the appropriate configuration to include the SATA driver in
the coreboot image and use the SATA device as the default kernel
location.

The only device currently supported is the NM10 AHCI controller,
the driver will have to be extended to allow for different PCI
device IDs.

BUG=chromium-os:19837
TEST=manual

. program the updated image (including vbexport extensions and AHCI
  driver fixes) on an Alex.
. restart the machine and enter `vboot_twostop' at u-boot prompt.

Observe ChromeOS booting all the way to login screen.

Change-Id: Ie952754a3b8ef07d5e7b4eb92c85abd825da6201
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/7233