Fix some bugs in the i8402 driver which show up when no controller is present
If no controller is present, the i8402 driver should return immediately and not
attempt to operate on the missing hardware.

In kbd_input_empty, the status register is checked every millisecond to see
whether the input buffer is empty, up to a timeout which is tracked by
decrimenting a counter each time the check is performed. The decrement is
performed with a postfix -- operator, and the value of the counter is checked
in place. That means that when the counter reaches zero and the loop
terminates, it will actually be decrimented one more time and become -1. That
value is returned as the return value of the function. That would give the
right answer if it wasn't for that extra decrement because a timeout would
indicate that the buffer never became empty.

This change fixes both of those bugs.

BUG=chrome-os-partner:5849
TEST=Built and booted on both x86-alex and stumpy. Used the built in keyboard
on x86-alex.

Change-Id: Id79d2067237952f95caca3f32d4928723d1fcdc1
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: http://gerrit.chromium.org/gerrit/7248