Restore ability to print before relocation.
The problem with printing appears to be due to the corrupted
contents of the global data structure (allocated at a fixed
address for the before relocation use). There is no need to place
the global data structure at a fixed address, it is enough to
have it on the stack before relocation, as it is used only by the
functions in init_sequence_f table, and then passed to
board_init_r() which copies its contents immediately.

The CONFIG_SYS_INIT_GD_ADDR definition is not needed anymore and
is being removed.

BUG=chromium-os:20142
TEST=manual
 . program the new image on Alex
 . restart the machine

 Observe the relocation address and u-boot banner printed on the
 console.

Change-Id: Ifcdcd0c9da1e0ba901f319ff9b7145909300b8a1
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/7373
Reviewed-by: Stefan Reinauer <reinauer@google.com>