Allow smaller write units in SPI flash drivers.
There is a problem with x86 systems using Intel ICH chips as the
SPI controller: whereas the SPI chips allow for 256 byte pages
for write transactions, the controllers allow for only 64 bytes
per transaction. This causes failures when the chip driver
invokes the controller driver write() function passing a 256 byte
page: only 64 bytes get written.

This problem is not easy to fix in u-boot, this change suggests a
scheme to allow the size override for certain SPI controllers.

BUG=chromium-os:20105
TEST=manual

. this change was tested combined with other modification, the
  ability to read and save the 16KB environment in a maconix SPI
  flash was demonstrated.

Change-Id: Ic42f294da0e6abd554107d6189cabd7269370b96
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/7522
Reviewed-by: Stefan Reinauer <reinauer@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>