CHROMIUM: harmony: Very basic harmony (AKA dev-board) support.
Previously, we had dev-board as a symlink to seaboard,
which was a terrible idea (AKA: it wouldn't work at all).
Now, I've made a copy of it and made a few small changes,
mostly making sure that the nand WP pin is right and taking
the keyboard out.

I won't say that it's pretty, but it's better than what was
there.

Lots of TODOs, including trying to resolve this with the DTS
file in ../harmony and actually trying to set these values
to something sane.

This is part of a batch of CLs:
- http://gerrit.chromium.org/gerrit/7838
  Hack it so that bundle_firmware forces one copy of the BCT for NAND.
- http://gerrit.chromium.org/gerrit/5207
  CHROMIUM: cbootimage: Add Bctcopy parameter
- http://gerrit.chromium.org/gerrit/7841
  Revert "Hack tegra2_dev-board to think NAND block size is 64K."
- http://gerrit.chromium.org/gerrit/7837
  Add some very basic NAND support to cros_write_firmware.
- http://gerrit.chromium.org/gerrit/7840
  CHROMIUM: harmony: Very basic harmony (AKA dev-board) support.

BUG=chromium-os:17465
TEST=Used cros_write_firmware to flash u-boot to NAND; rebooted and
saw new u-boot

Change-Id: Ie8e1a8c1bb600e48158bccfd4be226f12a41d3c4
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/7840