Don't spam POST80 codes with slow IO functions
This patch prevents u-boot from "spamming" random progress codes on
a port 80 "post card".
The previous version of this patch just removed the delays in the "slow"
IO functions, as they do not need to be slow, however, this patch is
less intrusive.

It uses another unused port that is often used by BIOSes (and the Linux
Kernel) for small delay timing purposes.

Signed-off-by: Stefan Reinauer <reinauer@google.com>

BUG=none
TEST=boot coreboot+u-boot, see the right post codes come up.

Change-Id: I9d1cc6de4472f5bb1f7b5ec6959273bfbfaa760b
Reviewed-on: http://gerrit.chromium.org/gerrit/7778
Reviewed-by: Simon Glass <sjg@chromium.org>
Commit-Ready: Stefan Reinauer <reinauer@google.com>
Tested-by: Stefan Reinauer <reinauer@google.com>