CHROMIUM: arm: Specify debugging serial port at boot.
This works together with a kernel change that looks at
the scratchpad register to determine which of the many
UARTs it should use for early printing:
<http://gerrit.chromium.org/gerrit/8148>.  While it is
unfortunate to need to pass this information in a
second way (it's already in the device tree), this
does allow the very early boot code (decompressing stub
and early assembly code) to print to the right port.

At the moment, I'm adding this to the UART init function.
Alternatively, we could add a more complex patch to
key off of the 'console' setting.

BUG=chromium-os:19656
TEST=See <http://gerrit.chromium.org/gerrit/8148>

Change-Id: Iea374625f22b9fa64cf22820dfd530eca1c38bb0
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/8156