CHROMIUM: tegra: mmc: Fixed handling of timeouts.
For some reason we are getting timeouts in the tegra MMC code
when we're reading from external MMC cards.  These don't seem
to be detrimental if we handle them properly.  This CL properly
clears the "normal interrupt status register" (norintsts) in
error conditions.  If we don't do this, when we come back into
mmc_send_cmd() the register will still contain status from the
last transaction.

BUG=chromium-os:20947
TEST=Booted from MMC 1
Also did the following for detailed test:
0. Ran with legacy u-boot and held space at bootup to get prompt.
1. mmc rescan 1
2. fatload mmc 1:c 0x408000 /u-boot/boot.scr.uimg ff
3. md.b 0x408000 0xff
4. Validated that the above looked like the bootscript (aka
   contained text like "boot-A.scr").

Change-Id: Ie4e7736d9b2055dcb4b28775545291b7eb3e633e
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/8469
Reviewed-by: Simon Glass <sjg@chromium.org>
Tested-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>