tegra: Build early code for ARMv4T
The Tegra family has an ARMv4T chip which runs board_init_f(). It does not
support the BLX instruction which the init_sequence function table generates.
The easiest fix is to build this file for ARMv4T on Tegra.

The alternative is to move the arch_cpu_init() call out of the init
sequence. But that fix would not make the reason for the change so obvious.

There is no code size impact.

BUG=chromium-os:19353
TEST=build and boot on Seaboard

Change-Id: Idb0ab17bd0839c472cda7bf4940522245792e328
Reviewed-on: http://gerrit.chromium.org/gerrit/8684
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>