Fix the inline assembly constraints in boot_zimage
It's possible that the compiler will decide to put the kernel entry point into
register %ebp which the assembly blob will set to 0 manually before it's used.
These new constraints will force it to use %eax for the kernel entry point,
and also tell it that the %ebp register is modified.

BUG=None
TEST=Built and booted on Stumpy and verified that the crash entering the kernel
no longer happens.

Change-Id: Iae321543e610f2084957978e438e91b1081dbc89
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: http://gerrit.chromium.org/gerrit/9764
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>