Reorder x86's post relocation memory layout
This changes the layout in decreasing addresses from:

1. Stack
2. Sections in the image
3. Heap

to

1. Sections in the image
2. Heap
3. Stack

This allows the stack to grow significantly more since it isn't constrained by
the other u-boot areas. More importantly, the generic memory wipe code assumes
that the stack is the lowest addressed area used by the main part of u-boot.
In the original layout, that means that u-boot tramples all over itself. In
the new layout, it works.

BUG=chrome-os-partner:6194
BUG=chrome-os-partner:6195
TEST=Booted on Stumpy.

Change-Id: I0484ad37ba15f63cb0fb289118cad360817ad11f
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: http://gerrit.chromium.org/gerrit/10006
Reviewed-by: Stefan Reinauer <reinauer@google.com>