Make u-boot load the recovery key in developer mode
The recovery key is needed to display recovery reason information when the tab
key is pressed. If it's not loaded, the data in the in-memory copy of the GBB
can be anything, and vboot may take a very long time to process a malformed
key. This workaround, provided by clchiou, ensures that the recovery key is
loaded when in developer mode so that when the tab key is pressed vboot doesn't
get stuck.

BUG=chrome-os-partner:6266
TEST=Booted into dev mode and pressed tab. Saw the recovery info and that the
system didn't hang and could continue to boot once Ctrl+D was pressed.

Change-Id: I2b85933d38d879a33d3ae369a8249a86af3e9caf
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: http://gerrit.chromium.org/gerrit/10008
Reviewed-by: Randall Spangler <rspangler@chromium.org>
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>
Reviewed-by: Stefan Reinauer <reinauer@google.com>