Import the glibc implementation of x86 memset into u-boot
I ran four iterations with the two implementations and used Vadim's CBMEM
infrastructure to measure the time they took. These are all in microseconds,
and the timestamp portion of the raw output of cbmem.py is included in the bug.
The new implementation is about twice as fast as the old.

Old:
1. 418,286
2. 418,302
3. 418,298
4. 418,290

New:
1. 184,800
2. 194,629
3. 194,188
4. 192,718

BUG=chrome-os-partner:6487
TEST=Booted on Stumpy.

Change-Id: Iba398929cbba395e10851d676ae9d356ae670f41
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: http://gerrit.chromium.org/gerrit/10284
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>
Commit-Ready: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>