CHROMIUM: tegra3: i2c: Move DVC semantics choice to device tree
On the tegra30, it appears that the "DVC" i2c controller has been
normalized and no longer requires special semantics for accessing it
(it has also just been renamed to "i2c5").  This change makes it so
that we don't pick DVC semantics based on the periperal ID, but
instead allow the device tree to specify.

BUG=chromium-os:21540
TEST=Compiled / booted on Kaen

Change-Id: Idfd96d1193b5ac267c61416544c63cc03dab396d
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/10279
Reviewed-by: Simon Glass <sjg@chromium.org>