Add support for specifying an initrd with the zboot command
This change finishes plumbing the initrd support built into the zboot
mechanism out to the command interface.

It also fixes a bug in the command declaration where the kernel size could
be passed as an optional second parameter but not enough arguments were
allowed.

BUG=chrome-os-partner:6715
TEST=Built and booted on a Lumpy with the U-Boot command line exposed. Used
the Linux kernel in the system partition and the initramfs used for recovery
to boot the system. Observed that the recovery scripts in the initramfs ran.
The hardest part about implementing this change was testing because it took
many attempts to build/find an initrd/initramfs and command line that would
actually work. I'm confident the problems I had were from not knowing how to
set up or use an initramfs properly, not from the U-Boot support itself.

Signed-off-by: Gabe Black <gabeblack@google.com>

Change-Id: Iccc5ff62e170e738dfb429a6b0ef9e27f0e7d8a9
Reviewed-on: https://gerrit.chromium.org/gerrit/12027
Tested-by: Gabe Black <gabeblack@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>