X86: Make the upper bound on relocated symbols closed instead of open
When relocating symbols, U-Boot only relocates the ones that fall within the
range of address that are actually being moved. This change makes the upper
bound of that region closed (less than or equal to) instead of open (less
than) so that the __bss_end symbol is processed as well.

BUG=None
TEST=Built and booted on Lumpy. Built on Kaen.

Signed-off-by: Gabe Black <gabeblack@google.com>

Change-Id: Ic1ee7402621dc266b3776ea5bdf3254ef51bc4ac
Reviewed-on: https://gerrit.chromium.org/gerrit/14148
Reviewed-by: Bill Richardson <wfrichar@chromium.org>
Commit-Ready: Gabe Black (Do Not Use) <gabeblack@google.com>
Tested-by: Gabe Black (Do Not Use) <gabeblack@google.com>