X86: Don't assume U-Boot extends to the end of memory when wiping
On ARM, U-Boot is put at the end of memory, and the code that records the
unused areas of memory to be wiped uses the end of RAM as the upper bound
for the region occupied by U-Boot. On x86, U-Boot may have been relocated to
somewhere lower than the end of RAM because it can normally only address a 32
bit address space. This change makes the wiping code calculate where U-Boot
actually ends on x86 so the areas with large addresses will be wiped properly.

BUG=chrome-os-partner:7579
TEST=Built and booted on Lumpy, Stumpy, and Kaen. On Lumpy, verified that the
area with physical addresses above 4GB was now in the list to be wiped.

Change-Id: Ib7e0818e85daf512b76f85f4e68f88230d301cea
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/14223