usb: workaround non-working keyboards.
If the USB keyboard is not answering properly the first request on its
interrupt endpoint, just skip it and try the next one.

This workarounds an issue with a wireless mouse dongle which presents
itself both as a keyboard and a mouse but has a non-functional keyboard
interface.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BUG=chrome-os-partner:7559
TEST=Boot a stumpy with the wireless mouse and the keyboard plugged in
such a way that the mouse enumerates first. Check we are booting
properly and the key strokes are working in the firmware.

Change-Id: I574f2a4bdb3bc7824b5bd747a237ef0b79dd83ac
Reviewed-on: https://gerrit.chromium.org/gerrit/14362
Commit-Ready: Vincent Palatin <vpalatin@chromium.org>
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Stefan Reinauer <reinauer@chromium.org>
(cherry picked from commit 012bbf0ce0301be2482857e3f03b481dd15c2340)
Reviewed-on: https://gerrit.chromium.org/gerrit/14379