gen: Make the SCSI command aware of READ CAPACITY (16) and how to use it
The generic SCSI code in U-Boot was only aware of the READ CAPACITY (10) SCSI
command which can detect the size of disks up to 2TB in size. If that size is
exceeded, it should then try the READ CAPACITY (16) command which returns a
64 bit block count value.

BUG=chrome-os-partner:8180
TEST=In conjunction with the next change, built and booted into ChromeOS on
the Emerald Lake 2 CRB with a 250 GB SSD. Did the same but forced the READ
CAPACITY (10) command to saturate and the code to fall back to READ CAPACITY
(16). Note that this code has not be tested with a real SCSI disk, just the
AHCI code pretending to be a SCSI disk as it historically has.

Change-Id: Ia0ee3fa1264649f25065658d5d368101d39ce614
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/18060
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>