gen: Get disk/part_dos.c to compile when CONFIG_SYS_64BIT_LBA is turned on
The code in this file tries to print an LBA using the format specifier
"%10ld", but when CONFIG_SYS_64BIT_LBA is turned on, the argument passed in
this position is a 64 bit value and isn't compatible. This change selects the
appropriate format specifier, the original or "10lld", based on whether
CONFIG_SYS_64BIT_LBA is turned on.

BUG=chrome-os-partner:8180
TEST=Built for Stumpy with BUILD_PART_FS_STUFF turned on and
CONFIG_SYS_64BIT_LBA enabled and disable, and saw the compile errors go away.

Change-Id: Ib80b1567ce07673aea2b47dfba9b3526b32f9de5
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/18219
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>