colibri_vf50: increase available space by booting from gfxram
Increase the available space for the U-Boot binary to half a megabyte
by booting from gfxRAM rather than sysRAM0.

Please note that according to the Vybrid Reference Manual the bigger
lower part of the sysRAM1 which immediately follows sysRAM0 should be
available as well but testing showed that as soon as the U-Boot binary
got bigger than 256 KB booting failed.