colibri_vf50: integrate config block handling
Integrate factory configuration block handling for hardware version,
MAC address and serial number to be passed to Linux.
The U-Boot environment is stored in NAND flash.

Default is to boot out of NAND flash using the UBI abstraction.
Optional NFS as well as MMC/SD boot variants are also available.

This brings the configuration more in-line with our Tegra based
versions.

While at it enable board size limit check set to 512 KB related to the
gfxRAM area of the OCRAM.

Please note that a boot delay of zero did not quite work the same way
as on the Tegras where one was still able to interrupt the boot process
by continuously hitting a key. Therefore set it to one second for now.