gen: usb: Fix EHCI periodic list handling
Our custom handling for periodic endpoints had some bugs in it:
allocating the wrong amount of memory, setting a bit only on low-speed
devices that needs to be present on all of them, and mixing up whether
to reenable the list after changing it between error and success cases.
This patch (hopefully) fixes all of them.

BRANCH=None
BUG=None
TEST=Make sure USB keyboard support in U-Boot works.

Change-Id: I0b302007a620927f596f30a48f420eb67371f78f
Signed-off-by: Julius Werner <jwe...@chromium.org>