apalis-imx6: use UARTs in DCE mode
The Apalis standart uses the UART in DTE mode.
This commit uses UART1 in DTE mode for the U-Boot console and
configures all used UARTs to start in DTE mode.

Note that for this to work module version V1.0A requires TXD/RXD to be crossed
between the Apalis iMX6 and the RS232 transceiver.