colibri vf50/61: fix environment offset
Environment offset was set to block 6 (0xC0000), which is still
inside the U-Boot partition. Since U-Boot is small enouth to fit
in the first 5 blocks, it usually is not an issue, however if
one of this 5 blocks is bad, then the environment overwrites part
of the boot loader. This fix sets the environment to erase block 12
(0x180000).