colibri_imx6.c: add patch_ddr_size cmd
Adds the patch_ddr_size cmd which patches the DCD data structure
settings for the DDR memory controller optimized for the module.
Currently this is only the bus width which is changed from 32bit
to 64bit on DL modules.

This allows to use a unified U-Boot for S and DL modules. Right after
flashing the U-Boot to eMMC this cmd will be run to complete the
update.