mtd: vf610_nfc: allow bitflips in an empty page
Allow bit flips in a empty page up to half of the recoverable
bits (strength / 2). Some flash show bit flips in empty pages
which are larger then the corrected bit count according to the
ECC controller. It is not yet clear how to solve this correctly,
discussion is ongoing:
http://thread.gmane.org/gmane.linux.ports.arm.kernel/295424

Since we habe a 24-bit correction, this allows up to 12 bit
flips on a empty page before reporting it as page with ECC
errors.