MLK-10774-41 imx: mx6sx: update VDDSOC standby voltage
This patch is from commit "f2c5102bf3763d77a227c1cba7fcd49e3db53a1d".
"
According the latest datasheet Rev.0,2/2015, the VDDSOC_IN voltage in standby/DSM
mode is 1.05V. As we use PFM mode of pFuse and this mode has 3% tolerance issue,
so the standby mode voltage should be (1.05 * 1.03) = 1.0815, we use 1.10V as the
minimal step is 25mV. For i.MX6sx SDB RevB boards, the VDDARM and VDDSOC use the
same supply, so the DSM voltage for VDDARM also need to be updated.
"

Signed-off-by: Peng Fan <Peng.Fan@freescale.com>