colibri_vf/imx6/7 apalis_imx6: synchronize/improve memory options
Synchronize initramfs related configuration options and improve
memory layout. The memory layout with an offset of 16MiB allows
to boot bigger kernels. With AUTO_ZRELADDR, which is pretty much
the standard nowadays, the kernel relacates itself to PC masked
with 0xf8000000 plus a text offset of 0x8000 (hence 0x80008000 for
Vybrid/i.MX 7 or 0x10008000 for i.MX 6). To avoid that the kernel
overwrites itself during uncompress, move the kernel further away
from its decompressed location. With a 16MiB offset from the start
of memory and a 16MiB distance to the device-tree, we allow kernel
up to ~16MiB.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>
Acked-by: Marcel Ziswiler <marcel.ziswiler@toradex.com>