update.sh: make set<media>update self contained
'run setupdate' tries a number of update media to find an update script
and then sources it.
Each media is probed by its own run set<media>update, however if a script
is found it is not sourced immediately but this task is delegated to
'run setupdate'. Thus one cannot just do a 'run setethupdate' to only
try to update over Ethernet.

This also unifies the functionality with Colibri VF and Colibri iMX7
were it has been implemented like this for a while now.

While at it use consistently ${loadaddr}.

Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>
Acked-by: Marcel Ziswiler <marcel.ziswiler@toradex.com>