tegra: lcd: video: add mem alloc for bmp colour conversion map
While trying to display a BMP image by running the command
'bmp display <address>' the Colibri T30 was getting stuck (while on
iMX6, iMX7 and Vybrid it worked). After investigation on what was going
wrong, I noticed that the function 'lcd_set_cmap' should get a pointer
address for the cmap variable, but the function returned 00000000.
After further investigation, I noticed that this value should be held
in the structure.member 'panel_info.cmap', which memory was being
dynamically allocated in the function 'lcd_ctrl_init' of the file
'drivers/video/sandbox_sdl.c' (which was not being called). After
copying the allocation to the same function in the file
'drivers/video/tegra.c', the command 'bmp display' didn't break
anymore. The Colibri T20 module was working without the patch, as the
module does have valid memory at the address 0, and of course it kept
working after the patch now properly allocating memory for it.

Signed-off-by: Leonardo Graboski Veiga <leonardo.veiga@toradex.com>
Signed-off-by: Marcel Ziswiler <marcel.ziswiler@toradex.com>