cmd_writebcb_mx7: create FCB without errata e9609 work-around
Create a FCB without work-around for errata e9609. With that FCB
in place the firmware can be written using the standard nand write
command like:
nand write ${loadaddr} u-boot ${filesize}

Using sectors_in_firmware{1|2} seems not to make any difference and
is also not set on Vybrid, hence do not set it anymore.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>
Acked-by: Marcel Ziswiler <marcel.ziswiler@toradex.com>