toradex: common: add common code which avoids loading environment
Add common code for i.MX SoC's which avoids loading the environment.
This reimplements the functionaltiy lost with ("common: autoboot:
do not reset environment") by not loading the environment from the
beginning. The main difference will be that dynamically set
environment variables such as "soc" in the i.MX 7 case will still
be set.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>
Acked-by: Marcel Ziswiler <marcel.ziswiler@toradex.com>