apalis-imx8: fix Ethernet
Make sure that all pins connected to the Micrel KSZ9031 PHY
are muxed. Properly reset the PHY after all muxing has been
applied. This makes sure that strapping is not overwritten by
the SoC default mux (particularly it makes sure that CLK125_NDO
is not driven low during reset).

Make sure to not use CONFIG_DM_ETH as it seems to break ETH
support as is.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>